refactor: search => shop으로 변경
